Extract thread actions initializer to its own file.
